Move back to singleton client
