+++ examples/with-apollo/lib/apollo.js
@@ -2,22 +2,51 @@ import React from 'react'
 import App from 'next/app'
 import Head from 'next/head'
 import { ApolloProvider } from '@apollo/react-hooks'
+import createApolloClient from '../apolloClient'
 
 let globalApolloClient = null
+
+// We expose this function so we are able to use
+// apolloClient inside getStaticProps, getStaticPaths or getServerProps
+export const ensureApolloClient = ctx => {
+  const inAppContext = Boolean(ctx.ctx)
+
+  if (process.env.NODE_ENV === 'development') {
+    if (inAppContext) {
+      console.warn(
+        'Warning: You have opted-out of Automatic Static Optimization due to `withApollo` in `pages/_app`.\n' +
+          'Read more: https://err.sh/next.js/opt-out-auto-static-optimization\n'
+      )
+    }
+  }
+
+  // Initialize ApolloClient if not already done
+  const apolloClient =
+    ctx.apolloClient || initApolloClient({}, inAppContext ? ctx.ctx : ctx)
+
+  // To avoid calling initApollo() twice in the server we send the Apollo Client as a prop
+  // to the component, otherwise the component would have to call initApollo() again but this
+  // time without the context, once that happens the following code will make sure we send
+  // the prop as `null` to the browser
+  apolloClient.toJSON = () => null
+
+  // Add apolloClient to NextPageContext & NextAppContext
+  // This allows us to consume the apolloClient inside our
+  // custom `getInitialProps({ apolloClient })`.
+  ctx.apolloClient = apolloClient
+  if (inAppContext) {
+    ctx.ctx.apolloClient = apolloClient
+  }
+
+  return ctx
+}
 
 /**
  * Creates a withApollo HOC
  * that provides the apolloContext
  * to a next.js Page or AppTree.
  */
-export const createWithApollo = createApolloClient => ({
-  ssr = true,
-} = {}) => PageComponent => {
-  if (typeof createApolloClient !== 'function') {
-    throw new Error(
-      '[withApollo] requires a function that returns an ApolloClient'
-    )
-  }
+export const withApollo = ({ ssr } = {}) => PageComponent => {
   const WithApollo = ({ apolloClient, apolloState, ...pageProps }) => {
     // Called by:
     // - getDataFromTree => apolloClient
@@ -27,8 +56,9 @@ export const createWithApollo = createApolloClient => ({
     if (apolloClient) {
       client = apolloClient
     } else {
-      client = initApolloClient(createApolloClient, apolloState, undefined)
+      client = initApolloClient(apolloState, undefined)
     }
+
     return (
       <ApolloProvider client={client}>
         <PageComponent {...pageProps} />
@@ -46,36 +76,8 @@ export const createWithApollo = createApolloClient => ({
 
   if (ssr || PageComponent.getInitialProps) {
     WithApollo.getInitialProps = async ctx => {
-      const { AppTree } = ctx
       const inAppContext = Boolean(ctx.ctx)
-
-      if (process.env.NODE_ENV === 'development') {
-        if (inAppContext) {
-          console.warn(
-            'Warning: You have opted-out of Automatic Static Optimization due to `withApollo` in `pages/_app`.\n' +
-              'Read more: https://err.sh/next.js/opt-out-auto-static-optimization\n'
-          )
-        }
-      }
-
-      if (ctx.apolloClient) {
-        throw new Error('Multiple instances of withApollo found.')
-      }
-
-      // Initialize ApolloClient
-      const apolloClient = initApolloClient(
-        createApolloClient,
-        {},
-        inAppContext ? ctx.ctx : ctx
-      )
-
-      // Add apolloClient to NextPageContext & NextAppContext
-      // This allows us to consume the apolloClient inside our
-      // custom `getInitialProps({ apolloClient })`.
-      ctx.apolloClient = apolloClient
-      if (inAppContext) {
-        ctx.ctx.apolloClient = apolloClient
-      }
+      const { apolloClient } = ensureApolloClient(ctx)
 
       // Run wrapped getInitialProps methods
       let pageProps = {}
@@ -87,14 +89,15 @@ export const createWithApollo = createApolloClient => ({
 
       // Only on the server:
       if (typeof window === 'undefined') {
+        const { AppTree } = ctx
         // When redirecting, the response is finished.
         // No point in continuing to render
         if (ctx.res && ctx.res.finished) {
           return pageProps
         }
 
-        // Only if ssr is enabled
-        if (ssr) {
+        // Only if dataFromTree is enabled
+        if (ssr && AppTree) {
           try {
             // Run all GraphQL queries
             const { getDataFromTree } = await import('@apollo/react-ssr')
@@ -127,16 +130,10 @@ export const createWithApollo = createApolloClient => ({
       // Extract query data from the Apollo store
       const apolloState = apolloClient.cache.extract()
 
-      // To avoid calling initApollo() twice in the server we send the Apollo Client as a prop
-      // to the component, otherwise the component would have to call initApollo() again but this
-      // time without the context, once that happens the following code will make sure we send
-      // the prop as `null` to the browser
-      apolloClient.toJSON = () => null
-
       return {
         ...pageProps,
         apolloState,
-        apolloClient,
+        apolloClient: ctx.apolloClient,
       }
     }
   }
@@ -149,7 +146,7 @@ export const createWithApollo = createApolloClient => ({
  * Creates or reuses apollo client in the browser.
  * @param  {Object} initialState
  */
-const initApolloClient = (createApolloClient, initialState, ctx) => {
+const initApolloClient = (initialState, ctx) => {
   // Make sure to create a new client for every server-side request so that data
   // isn't shared between connections (which would be bad)
   if (typeof window === 'undefined') {
