+++ examples/with-apollo/pages/index.js
@@ -3,7 +3,7 @@ import InfoBox from '../components/InfoBox'
 import Header from '../components/Header'
 import Submit from '../components/Submit'
 import PostList from '../components/PostList'
-import withApollo from '../lib/withApollo'
+import { withApollo } from '../lib/apollo'
 
 const IndexPage = props => (
   <App>
